requirements.txt
arrow==1.1.1
asgiref==3.4.0
binaryornot==0.4.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cookiecutter==1.7.3
Django==3.2.4
django-tailwind==2.0.1
djangorestframework==3.12.4
idna==2.10
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
poyo==0.5.0
psycopg2==2.9.1
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
